Modulize the app
